Add dockerfile for cuda-10.2
Rename cuda to cuda-11.2

+++ tools/docker/run_docker.sh
@@ -1,18 +1,25 @@
 #!/usr/bin/env bash
 set -e  # exit on error
 
-DOCKER_IMAGE=${DOCKER_IMAGE:-ava-cuda}
+if [ -z "$1" ]
+  then
+    echo "Please specify docker image name as the first argument"
+    echo "Usage $0 DOCKER_IMAGE_NAME"
+    exit 1
+fi
+
+DOCKER_IMAGE=${1}
+shift # Consume argument 1
 RUN_DOCKER_INTERACTIVE=${RUN_DOCKER_INTERACTIVE:-1}
 
 DEBUG_FLAGS="--cap-add=SYS_PTRACE --security-opt seccomp=unconfined"
-DOCKER_MAP="-v $PWD:$PWD -w $PWD -v /etc/passwd:/etc/passwd -v /etc/groups:/etc/groups"
+DOCKER_MAP="-v $PWD:$PWD -w $PWD -v /etc/passwd:/etc/passwd -v /etc/group:/etc/group"
 DOCKER_FLAGS="--rm ${DOCKER_MAP} -u`id -u`:`id -g` --ipc=host --security-opt seccomp=unconfined ${DEBUG_FLAGS}"
-if [ ${DOCKER_IMAGE} == "ava-rocm" ]; then
+if [[ ${DOCKER_IMAGE} == *"rocm"* ]]; then
     DOCKER_FLAGS="${DOCKER_FLAGS} --device=/dev/kfd --device=/dev/dri --group-add video"
-elif [ ${DOCKER_IMAGE} == "ava-cuda" ]; then
+elif [[ ${DOCKER_IMAGE} == *"cuda"* ]]; then
     DOCKER_FLAGS="${DOCKER_FLAGS} --gpus all"
 fi
-# -u `id -u`:`id -g`"
 
 if [ ${RUN_DOCKER_INTERACTIVE} -eq 1 ]; then
     DOCKER_CMD="docker run -it ${DOCKER_FLAGS} ${DOCKER_IMAGE}"
